ai/youtube_module/requirements.txt
 pydantic>=2.0.0
 python-dotenv>=1.0.0
 httpx>=0.24.0
 youtube-transcript-api>=0.6.1
 openai>=1.0.0
 rapidfuzz>=3.0.0

